% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_DmBTC.R
\name{Compare_DmBTC}
\alias{Compare_DmBTC}
\title{Run the DmBTC, Bamdit, and SCS Models}
\usage{
Compare_DmBTC(data)
}
\arguments{
\item{data}{The data to be used in the model.}
}
\value{
A list containing the results of the model fitting.
}
\description{
Run the DmBTC, Bamdit, and SCS Models
}
