% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DmBTCmeta.R
\name{DmBTCmeta}
\alias{DmBTCmeta}
\title{Run the DmBTC Model}
\usage{
DmBTCmeta(
  data,
  parameters,
  n.iter,
  n.burnin,
  n.chains,
  n.thin,
  digits,
  debug,
  inits = NULL
)
}
\arguments{
\item{data}{The data to be used in the model.}

\item{parameters}{The parameters to be estimated.}

\item{n.iter}{Number of iterations.}

\item{n.burnin}{Number of burn-in iterations.}

\item{n.chains}{Number of Markov chains.}

\item{n.thin}{Thinning interval.}

\item{digits}{Number of decimal places to display.}

\item{debug}{Whether to run in debug mode.}

\item{inits}{Initial values for the parameters.}
}
\value{
A list containing the results of the model fitting.
}
\description{
Run the DmBTC Model
}
