% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SROC.R
\name{plot_sroc}
\alias{plot_sroc}
\title{Plot the SROC plot}
\usage{
plot_sroc(data)
}
\arguments{
\item{data}{The output data from DmBTC model.}
}
\value{
A list containing the results of the model fitting.
}
\description{
Plot the SROC plot
}
