% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Traceplot_DmBTC.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot trace plots for MCMC chains}
\usage{
plot_trace(bugs.out, parameter = NULL)
}
\arguments{
\item{bugs.out}{The output object from \code{R2OpenBUGS::bugs()}.}

\item{parameter}{The parameter name to plot. If NULL, it will plot for all parameters.}
}
\value{
A trace plot for the MCMC chains.
}
\description{
This function plots trace plots for the MCMC chains generated by WinBUGS.
}
